maintaining two examples with logistic regression
